Refactor based on PR reviews
