+++ src/main/scala/tech/cryptonomic/conseil/tezos/DatabaseConversions.scala
@@ -1,7 +1,6 @@
 package tech.cryptonomic.conseil.tezos
 
 import tech.cryptonomic.conseil.tezos.TezosTypes._
-import tech.cryptonomic.conseil.model.Model
 import tech.cryptonomic.conseil.tezos.FeeOperations._
 import tech.cryptonomic.conseil.util.Conversion
 import tech.cryptonomic.conseil.util.Conversion.Id
